requirements.txt
fastapi
uvicorn
twilio
requests
python-dotenv
sqlalchemy
huggingface_hub
